StockDetailsFeatureTests
