+++ Wait/StockDetailsFeature/Datasource/StockOverviewDataSource.swift
@@ -9,6 +9,11 @@ import Logging
 import Model
 
 class StockOverviewDataSource: ObservableObject {
+
+  init(networkClient: StockOverviewNetworkClient) {
+    self.networkClient = networkClient
+  }
+
   // MARK: Internal
 
   @Published var stockOverview = StockOverview.empty
@@ -26,6 +31,6 @@ class StockOverviewDataSource: ObservableObject {
 
   // MARK: Private
 
-  private let networkClient = StockOverviewNetworkClient()
+  private let networkClient: StockOverviewNetworkClient
   private var subscriptions = [AnyCancellable]()
 }
