+++ Wait/StockDetailsFeature/View/StockStatsView.swift
