+++ Wait/StockDetailsFeatureTests/StockStatsViewSnapshotTests.swift
@@ -1,0 +1,45 @@
+//
+// Created by: kai_chen on 12/29/21.
+// Copyright © 2021 Wait. All rights reserved.
+//
+
+import XCTest
+@testable import StockDetailsFeature
+import SwiftUI
+import SnapshotTesting
+
+class StockStatsViewSnapshotTests: XCTestCase {
+
+// Has an error. Disable now
+/*
+ ld: warning: Could not find or use auto-linked library 'XCTestSwiftSupport'
+ ld: warning: Could not find or use auto-linked framework 'XCTest'
+ Undefined symbols for architecture x86_64:
+   "_OBJC_CLASS_$_XCTContext", referenced from:
+       objc-class-ref in SnapshotTesting.o
+   "_OBJC_CLASS_$_XCTAttachment", referenced from:
+       objc-class-ref in SnapshotTesting.o
+   "_OBJC_CLASS_$_XCTestExpectation", referenced from:
+       objc-class-ref in SnapshotTesting.o
+   "_OBJC_CLASS_$_XCTWaiter", referenced from:
+       objc-class-ref in SnapshotTesting.o
+   "__swift_FORCE_LOAD_$_XCTestSwiftSupport", referenced from:
+       __swift_FORCE_LOAD_$_XCTestSwiftSupport_$_SnapshotTesting in SnapshotTesting.o
+   "static (extension in XCTest):__C.XCTContext.runActivity<A>(named: Swift.String, block: (__C.XCTActivity) throws -> A) throws -> A", referenced from:
+       SnapshotTesting._verifyInlineSnapshot<A>(matching: @autoclosure () throws -> A, as: SnapshotTesting.Snapshotting<A, Swift.String>, record: Swift.Bool, timeout: Swift.Double, with: Swift.String, file: Swift.StaticString, testName: Swift.String, line: Swift.UInt) -> Swift.String? in SnapshotTesting.o
+       SnapshotTesting.verifySnapshot<A, B>(matching: @autoclosure () throws -> A, as: SnapshotTesting.Snapshotting<A, B>, named: Swift.String?, record: Swift.Bool, snapshotDirectory: Swift.String?, timeout: Swift.Double, file: Swift.StaticString, testName: Swift.String, line: Swift.UInt) -> Swift.String? in SnapshotTesting.o
+   "XCTest.XCTFail(_: Swift.String, file: Swift.StaticString, line: Swift.UInt) -> ()", referenced from:
+       SnapshotTesting._assertInlineSnapshot<A>(matching: @autoclosure () throws -> A, as: SnapshotTesting.Snapshotting<A, Swift.String>, record: Swift.Bool, timeout: Swift.Double, with: Swift.String, file: Swift.StaticString, testName: Swift.String, line: Swift.UInt) -> () in SnapshotTesting.o
+       SnapshotTesting.assertSnapshot<A, B>(matching: @autoclosure () throws -> A, as: SnapshotTesting.Snapshotting<A, B>, named: Swift.String?, record: Swift.Bool, timeout: Swift.Double, file: Swift.StaticString, testName: Swift.String, line: Swift.UInt) -> () in SnapshotTesting.o
+ ld: symbol(s) not found for architecture x86_64
+ clang: error: linker command failed with exit code 1 (use -v to see invocation)
+ */
+
+//  func testView() {
+//    let view = StockStatsView(title: "PE", value: "10.0")
+//    let vc = UIHostingController(rootView: view)
+//
+//    assertSnapshot(matching: vc, as: .image)
+//  }
+
+}
